Rename config.rs -> config_io.rs
